Two commits together from the official repo > my fork
Adds a minimal install of React

Although this doesn't do much, it pulls in the React library and starts
handing the render tasks over to React.

It caused some linter errors which meant that `npm run build-production`
failed, but the errors came about because of the ES6 import syntax. This
is a linter issue.

Contintue Reacting, add basic Flux store/actions

At 414 KB this is still bigger, but it has to get worse before it gets
better.

Splitting off the functionality is messy here because of the
entanglement between the data, the display, and the events. Once
everything can be modularized it should all make much more sense. Right
now, however, it's a mixture of React and d3 doing the same thing and
that's messy.

Eventually most of the stuff that `synteny-vis.js` seems to be intended
for will move into `<App />` as the top-of-the-hierarchy app logic.
Things like the histogram code don't belong there and they will need to
eventually move into the histogram module.

I have already shuffled most of the options config dialog code out of
`synteny-vis.js` and into `<OptionsDialog />`. This seemed to fit since
that form is all app-specific, but it builds the dialog form the
slightly-more-generic and mostly modular reincarnation of `makeForm()` -
`<Form />`.

As of now, I _believe_ that this is at feature-parity with **master**.